Mostly corrections for 64-bit FreeBSD. (plus one bug fix in endpoint.c)
io.c
    adding include of <string.h>

endpoint.c
    Fixing buf for skip record allocation. (had MIN where I should have had MAX)
    Changing type for 'waitfor' variable. - needed for 64-bit.

fts.c
    casting to remove warning on 64-bit

owampd.c,owampdP.h,owping.c,owpingP.h
    change type for 'waitfor'

jeff


git-svn-id: https://svn.internet2.edu/svn/owamp/trunk@880 a2d82226-602c-0410-bba7-e51b83dd10d2

+++ owampd/fts.c
@@ -1024,7 +1024,7 @@ fts_palloc(sp, more)
 	 * a signed int but in an FTSENT it is an unsigned short.
 	 * We limit fts_pathlen to USHRT_MAX to be safe in both cases.
 	 */
-	if (sp->fts_pathlen < 0 || sp->fts_pathlen >= USHRT_MAX) {
+	if (sp->fts_pathlen < 0 || sp->fts_pathlen >= (int)USHRT_MAX) {
 		if (sp->fts_path)
 			free(sp->fts_path);
 		sp->fts_path = NULL;
